test: extract model fixtures in a dedicated module
This would allow us to use those fixtures in tests of services.
