+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/test/java/fixtures/ApplicationFixtures.java
@@ -15,71 +15,8 @@
  */
 package fixtures;
 
-import io.gravitee.definition.model.Origin;
-import io.gravitee.rest.api.model.*;
-import io.gravitee.rest.api.model.application.ApplicationListItem;
-import io.gravitee.rest.api.model.application.ApplicationSettings;
-import io.gravitee.rest.api.model.application.OAuthClientSettings;
-import io.gravitee.rest.api.model.application.SimpleApplicationSettings;
-import java.util.Date;
-import java.util.Set;
-
 /**
  * @author Jeoffrey HAEYAERT (jeoffrey.haeyaert at graviteesource.com)
  * @author GraviteeSource Team
  */
-public class ApplicationFixtures {
-
-    protected static final ApplicationSettings.ApplicationSettingsBuilder APPLICATION_SETTINGS_BUILDER = ApplicationSettings
-        .builder()
-        .app(SimpleApplicationSettings.builder().clientId("clientId").build())
-        .oAuthClient(OAuthClientSettings.builder().clientId("clientId").build());
-
-    private static final ApplicationListItem.ApplicationListItemBuilder BASE_APPLICATION_LIST_ITEM = ApplicationListItem
-        .builder()
-        .id("my-application")
-        .name("My application")
-        .description("Description")
-        .apiKeyMode(ApiKeyMode.EXCLUSIVE)
-        .background("background")
-        .backgroundUrl("https://background.gravitee.io")
-        .createdAt(new Date())
-        .updatedAt(new Date())
-        .disableMembershipNotifications(true)
-        .domain("domain")
-        .groups(Set.of("group1", "group2"))
-        .origin(Origin.MANAGEMENT)
-        .picture("picture")
-        .pictureUrl("https://picture.gravitee.io")
-        .primaryOwner(PrimaryOwnerFixtures.aPrimaryOwnerEntity())
-        .settings(APPLICATION_SETTINGS_BUILDER.build())
-        .status("ACTIVE")
-        .type("iOS");
-
-    private static final ApplicationEntity.ApplicationEntityBuilder BASE_APPLICATION_ENTITY = ApplicationEntity
-        .builder()
-        .id("my-application")
-        .name("My application")
-        .description("Description")
-        .apiKeyMode(ApiKeyMode.EXCLUSIVE)
-        .background("background")
-        .createdAt(new Date())
-        .updatedAt(new Date())
-        .disableMembershipNotifications(true)
-        .domain("domain")
-        .groups(Set.of("group1", "group2"))
-        .origin(Origin.MANAGEMENT)
-        .picture("picture")
-        .primaryOwner(PrimaryOwnerFixtures.aPrimaryOwnerEntity())
-        .settings(APPLICATION_SETTINGS_BUILDER.build())
-        .status("ACTIVE")
-        .type("iOS");
-
-    public static ApplicationListItem anApplicationListItem() {
-        return BASE_APPLICATION_LIST_ITEM.build();
-    }
-
-    public static ApplicationEntity anApplicationEntity() {
-        return BASE_APPLICATION_ENTITY.build();
-    }
-}
+public class ApplicationFixtures extends ApplicationModelFixtures {}
