+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/test/java/fixtures/CorsFixtures.java
@@ -23,18 +23,7 @@ import java.util.Set;
  * @author GraviteeSource Team
  */
 @SuppressWarnings("ALL")
-public class CorsFixtures {
-
-    private static final io.gravitee.definition.model.Cors.CorsBuilder BASE_MODEL_CORS = io.gravitee.definition.model.Cors
-        .builder()
-        .accessControlAllowCredentials(true)
-        .accessControlAllowHeaders(Set.of("header1", "header2"))
-        .accessControlAllowMethods(Set.of("method1", "method2"))
-        .accessControlAllowOrigin(Set.of("origin1", "origin2"))
-        .enabled(true)
-        .accessControlExposeHeaders(Set.of("exposeHeader1", "exposeHeader2"))
-        .accessControlMaxAge(10)
-        .runPolicies(true);
+public class CorsFixtures extends CorsModelFixtures {
 
     private static final Cors.CorsBuilder BASE_CORS = Cors
         .builder()
@@ -46,10 +35,6 @@ public class CorsFixtures {
         .exposeHeaders(Set.of("exposeHeader1", "exposeHeader2"))
         .maxAge(10)
         .runPolicies(true);
-
-    public static io.gravitee.definition.model.Cors aModelCors() {
-        return BASE_MODEL_CORS.build();
-    }
 
     public static Cors aCors() {
         return BASE_CORS.build();
