+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/test/java/fixtures/EndpointFixtures.java
@@ -15,8 +15,6 @@
  */
 package fixtures;
 
-import io.gravitee.definition.model.v4.endpointgroup.Endpoint;
-import io.gravitee.definition.model.v4.endpointgroup.EndpointGroup;
 import io.gravitee.rest.api.management.v2.rest.model.EndpointGroupServices;
 import io.gravitee.rest.api.management.v2.rest.model.EndpointGroupServicesV2;
 import io.gravitee.rest.api.management.v2.rest.model.EndpointGroupV2;
@@ -32,38 +30,9 @@ import java.util.Collections;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Set;
 
 @SuppressWarnings("ALL")
-public class EndpointFixtures {
-
-    private static final io.gravitee.definition.model.Endpoint.EndpointBuilder BASE_MODEL_ENDPOINT_V2 =
-        io.gravitee.definition.model.Endpoint
-            .builder()
-            .name("Endpoint name")
-            .target("http://gravitee.io")
-            .weight(1)
-            .backup(false)
-            .status(io.gravitee.definition.model.Endpoint.Status.UP)
-            .tenants(List.of("tenant1", "tenant2"))
-            .type("HTTP")
-            .inherit(true)
-            .healthCheck(io.gravitee.definition.model.services.healthcheck.EndpointHealthCheckService.builder().build())
-            .configuration(null);
-
-    private static final io.gravitee.definition.model.EndpointGroup.EndpointGroupBuilder BASE_MODEL_ENDPOINTGROUP_V2 =
-        io.gravitee.definition.model.EndpointGroup
-            .builder()
-            .name("Endpoint group name")
-            .endpoints(Set.of(BASE_MODEL_ENDPOINT_V2.build()))
-            .loadBalancer(
-                io.gravitee.definition.model.LoadBalancer.builder().type(io.gravitee.definition.model.LoadBalancerType.ROUND_ROBIN).build()
-            )
-            .services(null)
-            .httpProxy(null)
-            .httpClientOptions(null)
-            .httpClientSslOptions(null)
-            .headers(Collections.emptyList());
+public class EndpointFixtures extends EndpointModelFixtures {
 
     private static final HttpEndpointV2.HttpEndpointV2Builder BASE_HTTP_ENDPOINT_V2 = HttpEndpointV2
         .builder()
@@ -94,32 +63,6 @@ public class EndpointFixtures {
         .httpClientOptions(null)
         .httpClientSslOptions(null);
 
-    private static final Endpoint.EndpointBuilder BASE_MODEL_ENDPOINT_V4 = Endpoint
-        .builder()
-        .name("Endpoint name")
-        .type("http-get")
-        .weight(1)
-        .inheritConfiguration(false)
-        .secondary(false)
-        .tenants(List.of("tenant1", "tenant2"))
-        .configuration("{\n  \"nice\" : \"configuration\"\n}")
-        .sharedConfigurationOverride("{\n  \"nice\" : \"configuration\"\n}")
-        .services(io.gravitee.definition.model.v4.endpointgroup.service.EndpointServices.builder().healthCheck(null).build());
-
-    private static final EndpointGroup.EndpointGroupBuilder BASE_MODEL_ENDPOINTGROUP_V4 = EndpointGroup
-        .builder()
-        .name("Endpoint group name")
-        .type("http-get")
-        .loadBalancer(
-            io.gravitee.definition.model.v4.endpointgroup.loadbalancer.LoadBalancer
-                .builder()
-                .type(io.gravitee.definition.model.v4.endpointgroup.loadbalancer.LoadBalancerType.ROUND_ROBIN)
-                .build()
-        )
-        .sharedConfiguration("{\n  \"nice\" : \"configuration\"\n}")
-        .endpoints(List.of(BASE_MODEL_ENDPOINT_V4.build()))
-        .services(io.gravitee.definition.model.v4.endpointgroup.service.EndpointGroupServices.builder().healthCheck(null).build());
-
     private static final EndpointGroupServices.EndpointGroupServicesBuilder BASE_ENDPOINTGROUP_SERVICES_V4 = EndpointGroupServices
         .builder()
         .discovery(null)
@@ -146,10 +89,6 @@ public class EndpointFixtures {
         .endpoints(List.of(BASE_ENDPOINT_V4.build()))
         .services(BASE_ENDPOINTGROUP_SERVICES_V4.build());
 
-    public static io.gravitee.definition.model.Endpoint aModelEndpointV2() {
-        return BASE_MODEL_ENDPOINT_V2.build();
-    }
-
     public static HttpEndpointV2 anHtpEndpointV2() {
         return BASE_HTTP_ENDPOINT_V2.build();
     }
@@ -158,24 +97,12 @@ public class EndpointFixtures {
         return new EndpointV2(anHtpEndpointV2());
     }
 
-    public static io.gravitee.definition.model.EndpointGroup aModelEndpointGroupV2() {
-        return BASE_MODEL_ENDPOINTGROUP_V2.build();
-    }
-
     public static EndpointGroupV2 anEndpointGroupV2() {
         return BASE_ENDPOINTGROUP_V2.build();
     }
 
-    public static Endpoint aModelEndpointV4() {
-        return BASE_MODEL_ENDPOINT_V4.build();
-    }
-
     public static EndpointV4 anEndpointV4() {
         return BASE_ENDPOINT_V4.build();
-    }
-
-    public static EndpointGroup aModelEndpointGroupV4() {
-        return BASE_MODEL_ENDPOINTGROUP_V4.build();
     }
 
     public static EndpointGroupV4 anEndpointGroupV4() {
