+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/test/java/fixtures/EntrypointFixtures.java
@@ -22,15 +22,7 @@ import java.util.LinkedHashMap;
 import java.util.Map;
 
 @SuppressWarnings("ALL")
-public class EntrypointFixtures {
-
-    private static final io.gravitee.definition.model.v4.listener.entrypoint.Entrypoint.EntrypointBuilder BASE_MODEL_ENTRYPOINT_V4 =
-        io.gravitee.definition.model.v4.listener.entrypoint.Entrypoint
-            .builder()
-            .type("Entrypoint type")
-            .qos(io.gravitee.definition.model.v4.listener.entrypoint.Qos.AT_LEAST_ONCE)
-            .dlq(new io.gravitee.definition.model.v4.listener.entrypoint.Dlq("my-endpoint"))
-            .configuration("{\"nice\": \"configuration\"}");
+public class EntrypointFixtures extends EntrypointModelFixtures {
 
     private static final Entrypoint.EntrypointBuilder BASE_ENTRYPOINT_V4 = Entrypoint
         .builder()
@@ -38,10 +30,6 @@ public class EntrypointFixtures {
         .qos(Qos.AT_LEAST_ONCE)
         .dlq(new Dlq().endpoint("my-endpoint"))
         .configuration(new LinkedHashMap<>(Map.of("nice", "configuration")));
-
-    public static io.gravitee.definition.model.v4.listener.entrypoint.Entrypoint aModelEntrypointV4() {
-        return BASE_MODEL_ENTRYPOINT_V4.build();
-    }
 
     public static Entrypoint anEntrypointV4() {
         return BASE_ENTRYPOINT_V4.build();
