+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/test/java/fixtures/FlowFixtures.java
@@ -15,10 +15,14 @@
  */
 package fixtures;
 
-import io.gravitee.definition.model.v4.flow.Flow;
-import io.gravitee.definition.model.v4.flow.selector.ChannelSelector;
-import io.gravitee.definition.model.v4.flow.step.Step;
-import io.gravitee.rest.api.management.v2.rest.model.*;
+import io.gravitee.rest.api.management.v2.rest.model.BaseSelector;
+import io.gravitee.rest.api.management.v2.rest.model.FlowV2;
+import io.gravitee.rest.api.management.v2.rest.model.FlowV4;
+import io.gravitee.rest.api.management.v2.rest.model.Operator;
+import io.gravitee.rest.api.management.v2.rest.model.PathOperator;
+import io.gravitee.rest.api.management.v2.rest.model.Selector;
+import io.gravitee.rest.api.management.v2.rest.model.StepV2;
+import io.gravitee.rest.api.management.v2.rest.model.StepV4;
 import java.util.LinkedHashMap;
 import java.util.List;
 import java.util.Map;
@@ -29,38 +33,20 @@ import java.util.Set;
  * @author GraviteeSource Team
  */
 @SuppressWarnings("ALL")
-public class FlowFixtures {
+public class FlowFixtures extends FlowModelFixtures {
 
-    private static final ChannelSelector.ChannelSelectorBuilder BASE_MODEL_CHANNEL_SELECTOR_V4 = ChannelSelector
+    private static final io.gravitee.rest.api.management.v2.rest.model.ChannelSelector.ChannelSelectorBuilder BASE_CHANNEL_SELECTOR_V4 = io.gravitee.rest.api.management.v2.rest.model.ChannelSelector
         .builder()
         .channel("channel")
+        .type(BaseSelector.TypeEnum.CHANNEL)
         .entrypoints(Set.of("entrypoint1", "entrypoint2"))
-        .operations(Set.of(ChannelSelector.Operation.SUBSCRIBE, ChannelSelector.Operation.PUBLISH))
-        .channelOperator(io.gravitee.definition.model.flow.Operator.EQUALS);
-
-    private static final io.gravitee.rest.api.management.v2.rest.model.ChannelSelector.ChannelSelectorBuilder BASE_CHANNEL_SELECTOR_V4 =
-        io.gravitee.rest.api.management.v2.rest.model.ChannelSelector
-            .builder()
-            .channel("channel")
-            .type(BaseSelector.TypeEnum.CHANNEL)
-            .entrypoints(Set.of("entrypoint1", "entrypoint2"))
-            .operations(
-                Set.of(
-                    io.gravitee.rest.api.management.v2.rest.model.ChannelSelector.OperationsEnum.SUBSCRIBE,
-                    io.gravitee.rest.api.management.v2.rest.model.ChannelSelector.OperationsEnum.PUBLISH
-                )
+        .operations(
+            Set.of(
+                io.gravitee.rest.api.management.v2.rest.model.ChannelSelector.OperationsEnum.SUBSCRIBE,
+                io.gravitee.rest.api.management.v2.rest.model.ChannelSelector.OperationsEnum.PUBLISH
             )
-            .channelOperator(io.gravitee.rest.api.management.v2.rest.model.Operator.EQUALS);
-
-    private static final Step.StepBuilder BASE_MODEL_STEP_V4 = Step
-        .builder()
-        .name("step")
-        .description("description")
-        .enabled(true)
-        .policy("policy")
-        .condition("{#context.attribute['condition'] == true}")
-        .messageCondition("{#context.attribute['messageCondition'] == true}")
-        .configuration("{\n  \"nice\" : \"config\"\n}");
+        )
+        .channelOperator(io.gravitee.rest.api.management.v2.rest.model.Operator.EQUALS);
 
     private static final StepV4.StepV4Builder BASE_STEP_V4 = StepV4
         .builder()
@@ -71,16 +57,6 @@ public class FlowFixtures {
         .condition("{#context.attribute['condition'] == true}")
         .messageCondition("{#context.attribute['messageCondition'] == true}")
         .configuration(new LinkedHashMap<>(Map.of("nice", "config")));
-
-    private static final Flow.FlowBuilder BASE_MODEL_FLOW_V4 = Flow
-        .builder()
-        .name("Flow")
-        .enabled(true)
-        .selectors(List.of(BASE_MODEL_CHANNEL_SELECTOR_V4.build()))
-        .request(List.of(BASE_MODEL_STEP_V4.name("step_request").build()))
-        .publish(List.of(BASE_MODEL_STEP_V4.name("step_publish").build()))
-        .response(List.of(BASE_MODEL_STEP_V4.name("step_response").build()))
-        .subscribe(List.of(BASE_MODEL_STEP_V4.name("step_subscribe").build()));
 
     private static final FlowV4.FlowV4Builder BASE_FLOW_V4 = FlowV4
         .builder()
@@ -93,15 +69,6 @@ public class FlowFixtures {
         .subscribe(List.of(BASE_STEP_V4.name("step_subscribe").build()))
         .tags(Set.of("tag1", "tag2"));
 
-    private static final io.gravitee.definition.model.flow.Step.StepBuilder BASE_MODEL_STEP_V2 = io.gravitee.definition.model.flow.Step
-        .builder()
-        .name("step")
-        .description("description")
-        .enabled(true)
-        .policy("policy")
-        .condition("{#context.attribute['condition'] == true}")
-        .configuration("{\n  \"nice\" : \"config\"\n}");
-
     private static final StepV2.StepV2Builder BASE_STEP_V2 = StepV2
         .builder()
         .name("step")
@@ -110,21 +77,6 @@ public class FlowFixtures {
         .policy("policy")
         .condition("{#context.attribute['condition'] == true}")
         .configuration(new LinkedHashMap<>(Map.of("nice", "config")));
-
-    private static final io.gravitee.definition.model.flow.Flow.FlowBuilder BASE_MODEL_FLOW_V2 = io.gravitee.definition.model.flow.Flow
-        .builder()
-        .name("Flow")
-        .enabled(true)
-        .pathOperator(
-            io.gravitee.definition.model.flow.PathOperator
-                .builder()
-                .operator(io.gravitee.definition.model.flow.Operator.EQUALS)
-                .path("/path")
-                .build()
-        )
-        .condition("{#context.attribute['condition'] == true}")
-        .pre(List.of(BASE_MODEL_STEP_V2.name("step_pre").build()))
-        .post(List.of(BASE_MODEL_STEP_V2.name("step_pot").build()));
 
     private static final FlowV2.FlowV2Builder BASE_FLOW_V2 = FlowV2
         .builder()
@@ -135,16 +87,8 @@ public class FlowFixtures {
         .pre(List.of(BASE_STEP_V2.name("step_pre").build()))
         .post(List.of(BASE_STEP_V2.name("step_pot").build()));
 
-    public static Flow aModelFlowV4() {
-        return BASE_MODEL_FLOW_V4.build();
-    }
-
     public static FlowV4 aFlowV4() {
         return BASE_FLOW_V4.build();
-    }
-
-    public static io.gravitee.definition.model.flow.Flow aModelFlowV2() {
-        return BASE_MODEL_FLOW_V2.build();
     }
 
     public static FlowV2 aFlowV2() {
