+++ gravitee-apim-rest-api/gravitee-apim-rest-api-management-v2/gravitee-apim-rest-api-management-v2-rest/src/test/java/fixtures/UserFixtures.java
@@ -15,18 +15,4 @@
  */
 package fixtures;
 
-import io.gravitee.rest.api.model.UserEntity;
-
-public class UserFixtures {
-
-    private static UserEntity.UserEntityBuilder USER_ENTITY = UserEntity
-        .builder()
-        .id("id")
-        .firstname("John")
-        .lastname("Doe")
-        .email("john@doe.com");
-
-    public static UserEntity aUserEntity() {
-        return USER_ENTITY.build();
-    }
-}
+public class UserFixtures extends UserModelFixtures {}
