+++ gravitee-apim-rest-api/gravitee-apim-rest-api-test-fixtures/pom.xml
@@ -1,0 +1,53 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<!--
+
+    Copyright © 2015 The Gravitee team (http://gravitee.io)
+
+    Licensed under the Apache License, Version 2.0 (the "License");
+    you may not use this file except in compliance with the License.
+    You may obtain a copy of the License at
+
+        http://www.apache.org/licenses/LICENSE-2.0
+
+    Unless required by applicable law or agreed to in writing, software
+    distributed under the License is distributed on an "AS IS" BASIS,
+    WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+    See the License for the specific language governing permissions and
+    limitations under the License.
+
+-->
+<project
+    xmlns="http://maven.apache.org/POM/4.0.0"
+    xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+    xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd"
+>
+	<modelVersion>4.0.0</modelVersion>
+
+	<parent>
+        <groupId>io.gravitee.apim.rest.api</groupId>
+        <artifactId>gravitee-apim-rest-api</artifactId>
+		<version>${revision}${sha1}${changelist}</version>
+    </parent>
+
+	<artifactId>gravitee-apim-rest-api-test-fixtures</artifactId>
+	<name>Gravitee.io APIM - Management API - Test Fixtures</name>
+
+
+	<properties>
+	</properties>
+
+	<dependencies>
+
+		<!-- Gravitee Management dependencies-->
+		<dependency>
+			<groupId>io.gravitee.apim.rest.api</groupId>
+			<artifactId>gravitee-apim-rest-api-model</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+		<dependency>
+			<groupId>io.gravitee.apim.rest.api</groupId>
+			<artifactId>gravitee-apim-rest-api-model</artifactId>
+			<version>${project.version}</version>
+		</dependency>
+	</dependencies>
+</project>
