+++ gravitee-apim-rest-api/pom.xml
@@ -47,6 +47,7 @@
         <module>gravitee-apim-rest-api-portal</module>
         <module>gravitee-apim-rest-api-standalone</module>
         <module>gravitee-apim-rest-api-spec-converter</module>
+        <module>gravitee-apim-rest-api-test-fixtures</module>
         <module>gravitee-apim-rest-api-coverage</module>
     </modules>
