+++ gravitee-apim-rest-api/sonar-project.properties
@@ -23,7 +23,7 @@ sonar.test.inclusions=**/*Test.java
 
 # Coverage
 sonar.coverage.jacoco.xmlReportPaths=gravitee-apim-rest-api-coverage/target/site/jacoco-aggregate/jacoco.xml
-sonar.coverage.exclusions=**/pom.xml, **/src/test/**, **/model/**, **/exception/**
+sonar.coverage.exclusions=**/pom.xml, **/src/test/**, **/model/**, **/exception/**, **/fixtures/**
 
 # Libraries
 sonar.libraries=${env.HOME}/.m2/repository/org/projectlombok/lombok/**/*.jar
